Add initial import - Linux kconfig clean fork
Every now and then a project is born, and they decide to use Linux's
kconfig to enable configuration of their project. As it stands we
*know* kconfig is now used in at least over 12 different projects [0].
kconfig then has become one of the leading industrial variability
modeling languages [1] [2].

What is often difficult to do is to start off using kconfig though
and integration into a project. Or updating / syncing to the latest
kconfig from upstream Linux. This is a relative passive fork which aims to
keep in sync with the Linux kernel's latest kconfig to make it easier to
keep up to date and to enable new projects to use and embrace kconfig on
their own. The goal is not to fork kconfig and evolve it separately,
but rather keep in sync with the evolution of kconfig on Linux to make
it easier for projects to use kconfig and also update their own kconfig
when needed.

Since we take code from Linux we respect its license and embrace
the GPLv2 for this project as well. We copy / embrace the SPDX license
conventions as well, and embrace the DCO (see CONTRIBUTING), and use
scripts/setlocalversion to help annotate versions from git if outside
of a signed tagged release.

[0] http://www.eng.uwaterloo.ca/~shshe/kconfig_semantics.pdf
[1] http://gsd.uwaterloo.ca/sites/default/files/vm-2013-berger.pdf
[2] http://gsd.uwaterloo.ca/sites/default/files/ase241-berger_0.pdf

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>