scripts/kconfig.Makefile: replace optional include with an non-dependency solution
Not all open source projects using kconfig include or require including
the generated .config file. As far as I can tell scripts don't look for
it on Linux -- I just checked and I cannot find an example.

However, including it on shell scripts is useful as you are essentially
then defining a set of variables, which can be used for a series of
things.

For projects which *do* want to take this approach, using an optional
include, on the scripts/kconfig.Makefile, ie:

 -include .config

will force that .config dependency to look for rules on your Makefile
targets, and run. For projects which have a .config rule on their
Makefiles which essentially just complains and tells the user that
they must first configure the project, this creates an issue, and
will *always* trigger that warning. Not because the user used a make
target which needed the .config, but just because scripts/kconfig.Makefile
uses an optional include.

We can work around this by open coding a non-dependency optional
include solution.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>